+++ app/src/main/java/CS2263/CourseProject/Subtask.java
@@ -3,15 +3,9 @@ package CS2263.CourseProject;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.List;
 
-public class Subtask {
-    @Getter @Setter private String priority;
-    @Getter @Setter private String title;
-    @Getter @Setter private int date;
-    @Getter @Setter private String description;
-    @Getter @Setter private String status;
-    @Getter @Setter private List<String> labels;
-   // @Getter @Setter private TaskList parentList;
+public class Subtask extends Task {
+    @Getter @Setter private Task parentTask;
+
 
 }