+++ app/src/main/java/CS2263/CourseProject/Task.java
@@ -16,5 +16,6 @@ public class Task {
     @Getter @Setter private String description;
     @Getter @Setter private String status;
     @Getter @Setter private List<String> labels;
+    @Getter @Setter private String priority;
    // @Getter @Setter private TaskList parentList;
 }